pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-contacts"
version = "0.1"
dependencies = [
    "django>=5.2",
    "django-localflavor~=4.0",
    "django-phonenumber-field[phonenumbers]~=8.1",
    "django-simple-history~3.8",
]
description = "A Django app providing contacts and their contact information."
readme = "README.rst"
requires-python = ">= 3.10"
authors = [
    {name = "Sandboxed Thoughts", email = "nate@sandboxedthoughts.com"},
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.1",  # Replace "X.Y" as appropriate
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.urls]
Homepage = "https://github.com/sandboxed-thoughts/django-contacts"
